Add some activities
