+++ app/src/main/java/com/lambency/lambency_client/Activities/ProfileActivity.java
@@ -1,0 +1,15 @@
+package com.lambency.lambency_client.Activities;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.lambency.lambency_client.R;
+
+public class ProfileActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_profile);
+    }
+}
